teste delete with null param throws exception

+++ section-9/rest-with-springboot/src/main/java/io/github/gabrielnavas/rest_with_springboot/person/PersonService.java
@@ -78,6 +78,10 @@ public class PersonService {
     }
 
     public void deletePerson(Long personId) {
+        if (personId == null) {
+            throw new IllegalArgumentException("missing person id param");
+        }
+        
         logger.info("Delete a Person!");
         Person person = personRepository.findById(personId).orElseThrow(() -> new PersonNotFoundException("id", String.format("%d", personId)));
         personRepository.delete(person);
